idk something

+++ src/battle_brain.h
@@ -1,5 +1,5 @@
 #pragma once
 #include "poke.h"
   
-int apply_damage (Pokemon * defender, Damage damage);
-Damage calculate_attack (Pokemon * attacker, Pokemon * defender, Move * move, int callback);
+uint8_t apply_damage (Pokemon * defender, Damage damage);
+Damage calculate_attack (Pokemon * attacker, Pokemon * defender, Move * move, uint8_t callback);
